prof: memoize EntityGridItem render with dumb component
